requirements.txt
certifi==2019.3.9
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.10.1
Werkzeug==0.15.2
wincertstore==0.2
Pillow==8.4.0